+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 
 import 'package:niku/niku.dart';
 
@@ -13,7 +14,9 @@ class MyApp extends StatelessWidget {
 }
 
 class HomePage extends StatelessWidget {
-  final baseText = NikuText("").color(Colors.white).fontSize(21);
+  final baseText = NikuText("")
+    ..color(Colors.white)
+    ..fontSize(21);
 
   build(context) {
     return Scaffold(
@@ -21,40 +24,48 @@ class HomePage extends StatelessWidget {
         body: NikuColumn([])
             .justifyCenter()
             .children([
-              NikuFlatButton(NikuText("Flat Button")
-                      .apply(baseText)
-                      .color(Colors.blue)
-                      .build())
+              NikuButton(
+                NikuText("Text Button")
+                    .apply(baseText)
+                    .color(Colors.blue)
+                    .build(),
+              )
                   .onPressed(() {})
-                  .splash(Colors.blueAccent.withOpacity(.125))
-                  .highlight(Colors.blueAccent.withOpacity(.125))
+                  .highlight(Colors.blueAccent.withOpacity(.075))
+                  .overlay(Colors.blueAccent.withOpacity(.125))
                   .rounded(8)
                   .px(40)
                   .py(20)
                   .my(20)
                   .build(),
-              NikuIconButton(Icon(Icons.image))
+              NikuIconButton(
+                Icon(Icons.image),
+              )
                   .setColor(
-                      color: Colors.blue,
-                      highlight: Colors.blueAccent.withOpacity(.1),
-                      splash: Colors.blueAccent.withOpacity(.2))
+                    color: Colors.blue,
+                    highlight: Colors.blueAccent.withOpacity(.1),
+                    splash: Colors.blueAccent.withOpacity(.2),
+                  )
                   .onPressed(() {})
                   .tooltip("Image")
                   .niku()
                   .my(8)
                   .build(),
-              NikuRaisedButton(NikuText("Raised Button").style(baseText))
+              NikuButton(
+                NikuText("Elevated Button").style(baseText),
+              )
                   .onPressed(() {})
                   .px(40)
                   .py(20)
                   .bg(Colors.blue)
-                  .splash(Colors.white.withOpacity(.25))
-                  .highlight(Colors.white.withOpacity(.125))
+                  .highlight(Colors.blue.shade500)
+                  .overlay(Colors.white.withOpacity(.25))
+                  .elevation(base: 2)
                   .rounded(8)
                   .my(20)
-                  .build(),
-              NikuOutlineButton(
-                NikuText("Outline Button")
+                  .elevated(),
+              NikuButton(
+                NikuText("Outlined Button")
                     .apply(baseText)
                     .color(Colors.blue)
                     .build(),
@@ -62,12 +73,14 @@ class HomePage extends StatelessWidget {
                   .onPressed(() {})
                   .px(40)
                   .py(20)
-                  .borderColor(Colors.black12)
-                  .splash(Colors.blueAccent.withOpacity(.25))
-                  .highlight(Colors.blueAccent.withOpacity(.125))
+                  .borderColor(
+                    base: Colors.black12,
+                    selected: Colors.blue,
+                  )
+                  .overlay(Colors.blueAccent.withOpacity(.125))
                   .rounded(8)
                   .my(20)
-                  .build(),
+                  .outlined(),
               NikuText("Material")
                   .fontSize(21)
                   .color(Colors.white)
@@ -84,32 +97,42 @@ class HomePage extends StatelessWidget {
                   .build(),
               NikuStack([
                 Text("Normal Text"),
-                Text("Text with Niku Property").niku().mt(40).build()
+                Text("Text with Niku Property").niku().mt(40).build(),
               ]).topCenter().niku().maxHeight(100).build()
             ])
-            .prepend(NikuText("Niku")
-                .fontSize(21)
-                .w600()
-                .color(Colors.blue[400]!)
-                .niku()
-                .mb(20)
-                .build())
-            .append(NikuTextField("Niku Text Field")
-                .allowSelectAll(false)
-                .prefixIcon(Icon(Icons.edit))
-                .suffixIcon(NikuIconButton(Icon(Icons.delete))
-                    .onPressed(() {})
-                    .highlight(Colors.transparent)
-                    .splash(Colors.transparent)
-                    .build())
-                .color(Colors.black87)
-                .border(OutlineInputBorder(borderSide: BorderSide(width: 2)))
-                .focusedBorder(OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.blue, width: 2)))
-                .niku()
-                .maxWidth(320)
-                .my(20)
-                .build())
+            .prepend(
+              NikuText("Niku")
+                  .fontSize(21)
+                  .w600()
+                  .color(Colors.blue.shade400)
+                  .niku()
+                  .mb(20)
+                  .build(),
+            )
+            .append(
+              NikuTextField("Niku Text Field")
+                  .allowSelectAll(false)
+                  .prefixIcon(
+                    Icon(Icons.edit),
+                  )
+                  .suffixIcon(
+                    NikuIconButton(
+                      Icon(Icons.delete),
+                    )
+                        .onPressed(() {})
+                        .highlight(Colors.transparent)
+                        .splash(Colors.transparent)
+                        .build(),
+                  )
+                  .color(Colors.black87)
+                  .border(OutlineInputBorder(borderSide: BorderSide(width: 2)))
+                  .focusedBorder(OutlineInputBorder(
+                      borderSide: BorderSide(color: Colors.blue, width: 2)))
+                  .niku()
+                  .maxWidth(320)
+                  .my(20)
+                  .build(),
+            )
             .niku()
             .fullWidth()
             .center()