+++ lib/niku.dart
@@ -3,6 +3,7 @@ library niku;
 export './widget/axis.dart';
 export './widget/base.dart';
 export './widget/button.dart';
+export './widget/iconButton.dart';
 export './widget/text.dart';
 export './widget/textField.dart';